+++ component/miniParser/index.json
@@ -1,0 +1,6 @@
+{
+  "component": true,
+  "usingComponents": {
+    "miniParser": "./index"
+  }
+}
