+++ component/miniParser/index.wxml
@@ -1,0 +1,19 @@
+<block wx:if="{{parserData}}">
+  <block wx:for="{{parserData}}">
+    <!-- 文字 -->
+    <text wx:if="{{item.name==='text'}}" class="miniParser-text">{{item.text}}</text>
+    <!-- 图片 -->
+    <image wx:elif="{{item.name==='img'}}" class="miniParser-image {{item.attrs.class || ''}}" style="{{item.attrs.style}}" src="{{item.attrs.src}}"></image>
+    <!-- 视频 -->
+    <video wx:elif="{{item.name==='video'}}" class="miniParser-video {{item.attrs.class || ''}}" style="{{item.attrs.style}}" src="{{item.attrs.src}}"></video>
+    <!-- 换行 -->
+    <text wx:elif="{{item.name==='br'}}" class="miniParser-br">\n</text>
+    <!-- 默认 -->
+    <view wx:else class="miniParser-view {{item.attrs.class || ''}}" style="{{item.attrs.style}}">
+      <miniParser wx:if="{{item.children}}" htmlData="{{item.children}}"></miniParser>
+    </view>
+  </block>
+</block>
+
+
+
