+++ index.html
@@ -5,6 +5,6 @@
     <title>miniParser</title>
 </head>
 <body>
-    <script src="./dist/index.js" type="module"></script>
+    <script src="./server/index.js" type="module"></script>
 </body>
 </html>
