+++ package.json
@@ -12,6 +12,7 @@
     "@rollup/plugin-typescript": "^8.3.0",
     "prettier": "^2.4.1",
     "rollup": "^2.59.0",
+    "rollup-plugin-delete": "^2.0.0",
     "rollup-plugin-livereload": "^2.0.5",
     "rollup-plugin-serve": "^1.1.0",
     "rollup-plugin-terser": "^7.0.2",
