+++ rollup.build.config.js
@@ -1,6 +1,7 @@
 import babel from "@rollup/plugin-babel";
 import typescript from "@rollup/plugin-typescript";
 import { terser } from "rollup-plugin-terser";
+import del from "rollup-plugin-delete";
 
 const config = {
   input: "src/main.ts",
@@ -11,6 +12,7 @@ const config = {
     sourcemap: false,
   },
   plugins: [
+    del({ targets: "dist/*" }),
     typescript(),
     babel({
       exclude: "node_modules/**",
