+++ src/const.ts
@@ -180,3 +180,13 @@ export const decodeRegexp: [RegExp, string][] = [
   [/\r\n/g, ""],
   [/\n/g, ""],
 ];
+
+export type attrsMapType = { [key: string]: string };
+
+export type jsonDataType = {
+  type: string;
+  name: string;
+  attrs?: { [key: string]: string };
+  genKey?: number;
+  children?: jsonDataType[];
+}[];
